test-slim Phase 3 cluster 3: rename/merge audit, admin, artifacts, bootstrap tests
Rename only (verbatim):
- test_phase4_matter_close_cascade.py -> test_matter_close.py
- test_phase5_audit_cost_columns.py -> test_audit_cost.py
- test_phase13b_audit_gap_fill.py -> test_audit_coverage.py
- test_phase16_doctor.py -> test_doctor.py

Merges (one test per distinct failure mode, shared fixtures):
- phase5 reconstruction + r2 fixes + phase14.5a filters -> test_audit_reconstruction.py
- phase11 admin role + phase13b users listing + phase14.5c admin recon -> test_admin_api.py
- phase6 matter artifacts + phase13b artifacts api -> test_matter_artifacts.py
- phase12 bootstrap admin + phase13b bootstrap state -> test_bootstrap_admin.py

Deduped: cursor-filter-applies-to-all-sources (same failure mode as the
R2 no-row-dropped pagination regression) and the phase5 viewed-row test
(its `returned` assertion folded into the unified-payload-shape test).
Governance coverage reorganised, not thinned: close-audit vs
test_lmf_close_and_export_audit.py and test_audit_coverage.py vs
test_provider_audit_completeness.py checked for overlap — no
byte-equivalent assertions, all kept.

Adds backend/tests/README.md (module -> test-file map, post-Phase-3
names) and updates AUDIT_COVERAGE_MATRIX.md references.

Sources 4,391 lines -> 3,722 test lines. Full suite: 811 passed,
3 known macOS sandbox rlimit failures, zero collection errors.
