+++ backend/tests/test_matter_artifacts.py
@@ -1,26 +1,22 @@
-"""Phase 13b A — artifact list/read endpoint tests.
+"""Matter artifacts — write_artifact substrate + list/read API.
 
-Six tests:
+Merged from test_phase6_matter_artifacts.py and
+test_phase13b_artifacts_api.py (test-slim Phase 3).
 
-1. Happy: list returns N rows in created_at desc order
-2. Happy: read returns payload + metadata
-3. Non-owner stranger: 404 uniform
-4. Archived matter: 404
-5. Artifact id not in this matter: 404 (defence-in-depth)
-6. Storage file missing on disk: 500 with structured error
-
-Plus a bonus reading test that NO audit row lands (Phase 13b
-Decision #1).
+Substrate: write_artifact creates object-storage file + row, the
+(invocation_id, kind) uniqueness constraint, and the WORM trigger
+(no UPDATE, no DELETE). API: list/read happy paths, uniform 404s
+(stranger / archived / cross-matter), 410 on missing storage object,
+and the no-read-audit decision.
 """
 
 from __future__ import annotations
 
 import json
 import uuid
-from pathlib import Path
 
 import pytest
-from sqlalchemy import select
+from sqlalchemy import select, text
 
 from app.core.matter_artifacts import write_artifact
 from app.models import (
@@ -28,6 +24,7 @@ from app.models import (
     Matter,
     MatterArtifact,
     PRIVILEGE_CLEARED,
+    PRIVILEGE_MIXED,
     STATUS_ARCHIVED,
     STATUS_OPEN,
     User,
@@ -35,29 +32,31 @@ from app.models import (
 
 
 # ---------------------------------------------------------------------------
-# Helpers
+# Shared helpers
 # ---------------------------------------------------------------------------
 
 
-async def _register_and_login(client, *, suffix: str = "") -> str:
-    email = f"p13ba{suffix}-{uuid.uuid4().hex[:8]}@example.com"
-    password = "phase13ba-2026"
-    await client.post(
-        "/auth/register", json={"email": email, "password": password}
+async def _make_user(db_session) -> User:
+    user = User(
+        id=uuid.uuid4(),
+        email=f"artifacts-{uuid.uuid4().hex[:8]}@example.com",
+        hashed_password="x" * 32,
+        is_active=True,
+        is_verified=True,
+        is_superuser=False,
     )
-    await client.post(
-        "/auth/login",
-        data={"username": email, "password": password},
-        headers={"Content-Type": "application/x-www-form-urlencoded"},
-    )
-    return email
+    db_session.add(user)
+    await db_session.flush()
+    return user
 
 
-async def _make_matter(session, user_id: uuid.UUID, *, posture: str = PRIVILEGE_CLEARED) -> Matter:
+async def _make_matter(
+    session, user_id: uuid.UUID, *, posture: str = PRIVILEGE_MIXED
+) -> Matter:
     m = Matter(
         id=uuid.uuid4(),
-        slug=f"p13ba-{uuid.uuid4().hex[:8]}",
-        title="P13b Artifact Test",
+        slug=f"art-{uuid.uuid4().hex[:8]}",
+        title="Artifact Test",
         matter_type="employment_tribunal",
         status=STATUS_OPEN,
         privilege_posture=posture,
@@ -67,6 +66,20 @@ async def _make_matter(session, user_id: uuid.UUID, *, posture: str = PRIVILEGE_
     session.add(m)
     await session.flush()
     return m
+
+
+async def _register_and_login(client, *, suffix: str = "") -> str:
+    email = f"artifacts{suffix}-{uuid.uuid4().hex[:8]}@example.com"
+    password = "artifacts-2026"
+    await client.post(
+        "/auth/register", json={"email": email, "password": password}
+    )
+    await client.post(
+        "/auth/login",
+        data={"username": email, "password": password},
+        headers={"Content-Type": "application/x-www-form-urlencoded"},
+    )
+    return email
 
 
 async def _seed_artifact(
@@ -90,9 +103,136 @@ async def _seed_artifact(
     )
 
 
-# ---------------------------------------------------------------------------
-# 1. Happy: list returns N rows in created_at desc order
-# ---------------------------------------------------------------------------
+# ===========================================================================
+# Substrate — write_artifact + constraints + WORM
+# ===========================================================================
+
+
+@pytest.mark.asyncio
+async def test_write_artifact_creates_file_and_row(db_session) -> None:
+    user = await _make_user(db_session)
+    matter = await _make_matter(db_session, user.id)
+    invocation_id = uuid.uuid4()
+    payload = {"findings": [{"clause": "x", "severity": "high"}]}
+
+    artifact = await write_artifact(
+        db_session,
+        matter=matter,
+        capability_id="examples.test.review",
+        module_id="examples.test",
+        invocation_id=invocation_id,
+        kind="findings_pack",
+        payload=payload,
+        actor_user_id=user.id,
+    )
+    await db_session.commit()
+
+    # Row populated.
+    assert artifact.id is not None
+    assert artifact.size_bytes > 0
+    assert artifact.kind == "findings_pack"
+    assert artifact.invocation_id == invocation_id
+
+    # Object in storage + parses as expected (LMF-1: artifacts are in
+    # object storage; storage_path is now an S3 key, not an fs path).
+    from app.core.storage import get_storage_backend
+    assert not artifact.storage_path.startswith("/")  # a key, not a path
+    data = get_storage_backend().get_bytes(artifact.storage_path)
+    assert json.loads(data.decode("utf-8")) == payload
+
+
+@pytest.mark.asyncio
+async def test_unique_invocation_kind_constraint(db_session) -> None:
+    user = await _make_user(db_session)
+    matter = await _make_matter(db_session, user.id)
+    invocation_id = uuid.uuid4()
+
+    await _seed_artifact(
+        session=db_session,
+        matter=matter,
+        user_id=user.id,
+        invocation_id=invocation_id,
+        payload={"a": 1},
+    )
+    await db_session.commit()
+
+    # Same (invocation_id, kind) → IntegrityError.
+    with pytest.raises(Exception):
+        await _seed_artifact(
+            session=db_session,
+            matter=matter,
+            user_id=user.id,
+            invocation_id=invocation_id,
+            payload={"a": 2},
+        )
+        await db_session.commit()
+    await db_session.rollback()
+
+
+@pytest.mark.asyncio
+async def test_different_kinds_on_same_invocation_allowed(db_session) -> None:
+    """One invocation can produce multiple artifact KINDS."""
+    user = await _make_user(db_session)
+    matter = await _make_matter(db_session, user.id)
+    invocation_id = uuid.uuid4()
+
+    a1 = await _seed_artifact(
+        session=db_session,
+        matter=matter,
+        user_id=user.id,
+        invocation_id=invocation_id,
+        kind="findings_pack",
+        payload={"a": 1},
+    )
+    a2 = await _seed_artifact(
+        session=db_session,
+        matter=matter,
+        user_id=user.id,
+        invocation_id=invocation_id,
+        kind="citation_pack",
+        payload={"b": 2},
+    )
+    await db_session.commit()
+    assert a1.id != a2.id
+
+
+@pytest.mark.asyncio
+async def test_worm_trigger_rejects_update(db_session) -> None:
+    user = await _make_user(db_session)
+    matter = await _make_matter(db_session, user.id)
+    artifact = await _seed_artifact(
+        session=db_session, matter=matter, user_id=user.id, payload={"a": 1}
+    )
+    await db_session.commit()
+    with pytest.raises(Exception):
+        await db_session.execute(
+            text("UPDATE matter_artifacts SET kind = 'tampered' WHERE id = :rid"),
+            {"rid": artifact.id},
+        )
+        await db_session.commit()
+    await db_session.rollback()
+
+
+@pytest.mark.asyncio
+async def test_worm_trigger_rejects_delete(db_session) -> None:
+    user = await _make_user(db_session)
+    matter = await _make_matter(db_session, user.id)
+    artifact = await _seed_artifact(
+        session=db_session, matter=matter, user_id=user.id, payload={"a": 1}
+    )
+    await db_session.commit()
+    with pytest.raises(Exception):
+        await db_session.execute(
+            text("DELETE FROM matter_artifacts WHERE id = :rid"),
+            {"rid": artifact.id},
+        )
+        await db_session.commit()
+    await db_session.rollback()
+
+
+# ===========================================================================
+# API — list/read endpoints
+# ===========================================================================
 
 
 @pytest.mark.asyncio
@@ -103,7 +243,7 @@ async def test_list_artifacts_returns_rows_in_desc_order(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         user = await session.scalar(select(User).where(User.email == email))
-        matter = await _make_matter(session, user.id)
+        matter = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         # Three artifacts on this matter.
         for _ in range(3):
@@ -130,11 +270,6 @@ async def test_list_artifacts_returns_rows_in_desc_order(client) -> None:
         assert "payload" not in row  # list returns summary only
 
 
-# ---------------------------------------------------------------------------
-# 2. Happy: read returns payload + metadata
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_read_artifact_returns_payload(client) -> None:
     email = await _register_and_login(client)
@@ -143,7 +278,7 @@ async def test_read_artifact_returns_payload(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         user = await session.scalar(select(User).where(User.email == email))
-        matter = await _make_matter(session, user.id)
+        matter = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         artifact = await _seed_artifact(
             session,
@@ -163,11 +298,6 @@ async def test_read_artifact_returns_payload(client) -> None:
     assert body["payload"]["findings"][0]["clause"] == "5.2"
 
 
-# ---------------------------------------------------------------------------
-# 3. Non-owner stranger: 404 uniform
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_non_owner_stranger_404(client) -> None:
     """Owner creates a matter with an artifact; stranger logs in and
@@ -178,7 +308,7 @@ async def test_non_owner_stranger_404(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         owner = await session.scalar(select(User).where(User.email == owner_email))
-        matter = await _make_matter(session, owner.id)
+        matter = await _make_matter(session, owner.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         artifact = await _seed_artifact(
             session, matter=matter, user_id=owner.id
@@ -199,11 +329,6 @@ async def test_non_owner_stranger_404(client) -> None:
     assert resp_read.status_code == 404
 
 
-# ---------------------------------------------------------------------------
-# 4. Archived matter: 404
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_archived_matter_404(client) -> None:
     email = await _register_and_login(client)
@@ -212,7 +337,7 @@ async def test_archived_matter_404(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         user = await session.scalar(select(User).where(User.email == email))
-        matter = await _make_matter(session, user.id)
+        matter = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         artifact = await _seed_artifact(session, matter=matter, user_id=user.id)
         matter.status = STATUS_ARCHIVED
@@ -229,11 +354,6 @@ async def test_archived_matter_404(client) -> None:
     assert resp_read.status_code == 404
 
 
-# ---------------------------------------------------------------------------
-# 5. Artifact id not in this matter: 404
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_artifact_from_other_matter_returns_404(client) -> None:
     """Defence-in-depth — the FK already enforces matter_id, but the
@@ -244,8 +364,8 @@ async def test_artifact_from_other_matter_returns_404(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         user = await session.scalar(select(User).where(User.email == email))
-        matter_a = await _make_matter(session, user.id)
-        matter_b = await _make_matter(session, user.id)
+        matter_a = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
+        matter_b = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         # Artifact lives on B.
         artifact_b = await _seed_artifact(
@@ -260,11 +380,6 @@ async def test_artifact_from_other_matter_returns_404(client) -> None:
     assert resp.status_code == 404
 
 
-# ---------------------------------------------------------------------------
-# 6. Storage file missing on disk: 500 with structured error
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_storage_object_missing_returns_410(client) -> None:
     email = await _register_and_login(client)
@@ -273,7 +388,7 @@ async def test_storage_object_missing_returns_410(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         user = await session.scalar(select(User).where(User.email == email))
-        matter = await _make_matter(session, user.id)
+        matter = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         artifact = await _seed_artifact(session, matter=matter, user_id=user.id)
         await session.commit()
@@ -293,11 +408,6 @@ async def test_storage_object_missing_returns_410(client) -> None:
     assert detail["artifact_id"] == str(artifact_id)
 
 
-# ---------------------------------------------------------------------------
-# Bonus: NO read audit (Phase 13b Decision #1)
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_artifact_read_emits_no_audit_row(client) -> None:
     """Decision #1: artifact reads do NOT emit an audit row."""
@@ -307,7 +417,7 @@ async def test_artifact_read_emits_no_audit_row(client) -> None:
     factory = app.state.session_factory
     async with factory() as session:
         user = await session.scalar(select(User).where(User.email == email))
-        matter = await _make_matter(session, user.id)
+        matter = await _make_matter(session, user.id, posture=PRIVILEGE_CLEARED)
         await session.flush()
         artifact = await _seed_artifact(session, matter=matter, user_id=user.id)
         await session.commit()
